feat: upgrade project to Vite and React 19, refactor structure and dependencies
- Changed project name and version in package.json
- Updated React and React DOM to version 19.1.0
- Added Vite as the build tool with corresponding scripts
- Refactored import paths to use alias '@' for cleaner imports
- Removed routes.js and migrated routes to routes.jsx
- Updated postcss.config.js to use ES module syntax
- Created index.html and main.jsx for Vite setup
- Added SVG logo to public directory
- Updated styles in index.css for consistency
- Adjusted sidebar and layout components to reflect new import paths
