+++ index.html
@@ -1,0 +1,43 @@
+<!--
+  _   _  ___  ____  ___ ________  _   _     _   _ ___   
+ | | | |/ _ \|  _ \|_ _|__  / _ \| \ | |   | | | |_ _| 
+ | |_| | | | | |_) || |  / / | | |  \| |   | | | || | 
+ |  _  | |_| |  _ < | | / /| |_| | |\  |   | |_| || |
+ |_| |_|\___/|_| \_\___/____\___/|_| \_|    \___/|___|
+                                                                                                                                                                                                                                                                                                                                       
+=========================================================
+* Horizon UI Tailwind React - v1.0.0
+=========================================================
+
+* Product Page: https://www.horizon-ui.com/
+* Copyright 2023 Horizon UI (https://www.horizon-ui.com/)
+
+* Designed and Coded by Simmmple
+
+=========================================================
+
+* The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
+
+-->
+
+<!doctype html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <link rel="icon" href="/favicon.ico"/>
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+        <meta name="theme-color" content="#000000"/>
+    <title>Horizon UI - Tailwind React</title>
+  </head>
+  <body class="font-dm bg-lightPrimary" >
+    <div id="root"></div>
+    <script type="module" src="/src/main.jsx"></script>
+  </body>
+</html>
+
+ <!--
+              ___
+          .__(  .)< 
+           \_____)
+       ~~~~~~~~~~~~~~~~    
+-->
