+++ package.json
@@ -1,7 +1,14 @@
 {
-  "name": "horizon-ui-tailwind-react",
-  "version": "2.0.0",
+  "name": "horizon-tailwind-react-vite",
   "private": true,
+  "version": "3.0.0",
+  "type": "module",
+  "scripts": {
+    "dev": "vite",
+    "build": "vite build",
+    "lint": "eslint .",
+    "preview": "vite preview"
+  },
   "dependencies": {
     "@chakra-ui/hooks": "^2.1.4",
     "@chakra-ui/modal": "^2.2.9",
@@ -16,45 +23,28 @@
     "@testing-library/react": "^16.0.0",
     "@testing-library/user-event": "^14.5.2",
     "apexcharts": "3.35.5",
-    "react": "19.0.0",
+    "react": "^19.1.0",
     "react-apexcharts": "1.4.0",
     "react-calendar": "^5.0.0",
-    "react-dom": "19.0.0",
+    "react-dom": "^19.1.0",
     "react-icons": "^4.4.0",
-    "react-router-dom": "^6.4.0",
-    "react-scripts": "5.0.1",
-    "tailwindcss-rtl": "^0.9.0",
-    "web-vitals": "^2.1.4"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "pretty": "prettier --write \"./**/*.{js,jsx,json}\""
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+    "react-router-dom": "6.4",
+    "tailwindcss-rtl": "^0.9.0"
   },
   "devDependencies": {
+    "@eslint/js": "^9.25.0",
+    "@tailwindcss/postcss": "^4.1.10",
+    "@tailwindcss/vite": "^4.1.10",
+    "@types/react": "^19.1.2",
+    "@types/react-dom": "^19.1.2",
+    "@vitejs/plugin-react": "^4.4.1",
+    "@vitejs/plugin-react-swc": "^3.9.0",
     "autoprefixer": "^10.4.8",
-    "postcss": "^8.4.16",
-    "prettier": "^2.8.3",
+    "eslint": "^9.25.0",
+    "eslint-plugin-react-hooks": "^5.2.0",
+    "eslint-plugin-react-refresh": "^0.4.19",
+    "globals": "^16.0.0",
+    "vite": "^6.3.5",
     "prettier-plugin-tailwindcss": "^0.2.1",
     "tailwindcss": "^3.1.8"
   }
