+++ src/layouts/admin/index.jsx
@@ -3,7 +3,7 @@ import { Routes, Route, Navigate, useLocation } from "react-router-dom";
 import Navbar from "components/navbar";
 import Sidebar from "components/sidebar";
 import Footer from "components/footer/Footer";
-import routes from "routes.js";
+import routes from "@/routes";
 
 export default function Admin(props) {
   const { ...rest } = props;
