+++ src/layouts/rtl/index.jsx
@@ -3,7 +3,7 @@ import { Routes, Route, Navigate, useLocation } from "react-router-dom";
 import Navbar from "components/navbar/RTL";
 import Sidebar from "components/sidebar/RTL";
 import Footer from "components/footer/Footer";
-import routes from "routes.js";
+import routes from "@/routes";
 
 export default function RTL(props) {
   const { ...rest } = props;
